Allow overriding the SMTP envelope from
The SMTP `MAIL FROM` command can specify a different address than the
one on the `From:` header. This address is called the 'envelope from'
and is used by MTAs for reporting bounced emails, amongst others.

This patch allows overriding the envelope from address by putting an
email address in the `:envelope_from` key of the `private` data of the
Bamboo email struct.