Add tls option for server_name_indication
OTP 26 is much more picky about TLS handshaking. In OTP 25, I was able
to send email over STARTTLS connections with only debug-level warnings,
but connections fail with `{:error, {:temporary_failure,
~c"52.192.253.213", :tls_failed}}` when running OTP 26.

In principle, I think it would be a bit less fussy to accept a configuration key
`tls_options` and pass those all through to GenSmtp, as is common in other
libraries (Mint, Slipstream do this if I recall) but this PR matches
existing conventions for this library.

I've added a couple of tests. For convenience I also automatically
convert binary to charlist, since server_name_indication must be a
charlist.